Revert "feat!: Support fractional grid and font sizes (#2485)" (#2497)
This reverts commit 12a415aebe02ad49a11486ec789aa438b3f77e61.

+++ src/editor/mod.rs
@@ -554,7 +554,7 @@ impl Editor {
             }
             GuiOption::LineSpace(linespace) => {
                 self.draw_command_batcher
-                    .queue(DrawCommand::LineSpaceChanged(linespace as f32));
+                    .queue(DrawCommand::LineSpaceChanged(linespace));
 
                 self.redraw_screen();
             }
